+++ packages/cojson/src/exports.ts
@@ -13,8 +13,9 @@ import {
   RawProfile,
   accountHeaderForInitialAgentSecret,
 } from "./coValues/account.js";
-import { RawCoList } from "./coValues/coList.js";
+import { OpID, RawCoList } from "./coValues/coList.js";
 import { RawCoMap } from "./coValues/coMap.js";
+import { RawCoPlainText, stringifyOpID } from "./coValues/coPlainText.js";
 import {
   CoStreamItem,
   RawBinaryCoStream,
@@ -136,6 +137,8 @@ export {
   emptyKnownState,
   emptyDataMessage,
   unknownDataMessage,
+  RawCoPlainText,
+  stringifyOpID,
 };
 
 export type {
@@ -150,6 +153,7 @@ export type {
   CoValueUniqueness,
   Stringified,
   CoStreamItem,
+  OpID,
 };
 
 // eslint-disable-next-line @typescript-eslint/no-namespace