+++ packages/cojson/src/permissions.ts
@@ -137,6 +137,7 @@ function resolveMemberStateFromParentReference(
   coValue: CoValueCore,
   memberState: MemberState,
   parentReference: ParentGroupReference,
+  extendChain: Set<CoValueCore["id"]>,
 ) {
   const parentGroup = coValue.node.expectCoValueLoaded(
     getParentGroupId(parentReference),
@@ -147,14 +148,21 @@ function resolveMemberStateFromParentReference(
     return;
   }
 
+  // Skip circular references
+  if (extendChain.has(parentGroup.id)) {
+    return;
+  }
+
   const initialAdmin = parentGroup.header.ruleset.initialAdmin;
 
   if (!initialAdmin) {
     throw new Error("Group must have initialAdmin");
   }
 
+  extendChain.add(parentGroup.id);
+
   const { memberState: parentGroupMemberState } =
-    determineValidTransactionsForGroup(parentGroup, initialAdmin);
+    determineValidTransactionsForGroup(parentGroup, initialAdmin, extendChain);
 
   for (const agent of Object.keys(parentGroupMemberState) as Array<
     keyof MemberState
@@ -171,6 +179,7 @@ function resolveMemberStateFromParentReference(
 function determineValidTransactionsForGroup(
   coValue: CoValueCore,
   initialAdmin: RawAccountID | AgentID,
+  extendChain?: Set<CoValueCore["id"]>,
 ): { validTransactions: ValidTransactionsResult[]; memberState: MemberState } {
   const allTransactionsSorted: {
     sessionID: SessionID;
@@ -305,7 +314,24 @@ function determineValidTransactionsForGroup(
         logPermissionError("Only admins can set parent extensions");
         continue;
       }
-      resolveMemberStateFromParentReference(coValue, memberState, change.key);
+
+      extendChain = extendChain ?? new Set([]);
+
+      resolveMemberStateFromParentReference(
+        coValue,
+        memberState,
+        change.key,
+        extendChain,
+      );
+
+      // Circular reference detected, drop all the transactions involved
+      if (extendChain.has(coValue.id)) {
+        logPermissionError(
+          "Circular extend detected, dropping the transaction",
+        );
+        continue;
+      }
+
       validTransactions.push({ txID: { sessionID, txIndex }, tx });
       continue;
     } else if (isChildExtension(change.key)) {
@@ -320,6 +346,7 @@ function determineValidTransactionsForGroup(
         );
         continue;
       }
+
       validTransactions.push({ txID: { sessionID, txIndex }, tx });
       continue;
     } else if (isWriteKeyForMember(change.key)) {