+++ packages/jazz-react-auth-clerk/CHANGELOG.md
@@ -1,5 +1,47 @@
 # jazz-browser-media-images
 
+## 0.9.1
+
+### Patch Changes
+
+- Updated dependencies [1b71969]
+  - jazz-react@0.9.1
+  - jazz-tools@0.9.1
+  - jazz-browser-auth-clerk@0.9.1
+
+## 0.9.0
+
+### Patch Changes
+
+- Updated dependencies [8eda792]
+- Updated dependencies [956a4d1]
+- Updated dependencies [8eda792]
+- Updated dependencies [1ef3226]
+  - cojson@0.9.0
+  - jazz-react@0.9.0
+  - jazz-tools@0.9.0
+  - jazz-browser-auth-clerk@0.9.0
+
+## 0.8.51
+
+### Patch Changes
+
+- Updated dependencies [dc62b95]
+- Updated dependencies [1de26f8]
+  - jazz-tools@0.8.51
+  - jazz-browser-auth-clerk@0.8.51
+  - jazz-react@0.8.51
+
+## 0.8.50
+
+### Patch Changes
+
+- Updated dependencies [43378ef]
+  - cojson@0.8.50
+  - jazz-browser-auth-clerk@0.8.50
+  - jazz-react@0.8.50
+  - jazz-tools@0.8.50
+
 ## 0.8.49
 
 ### Patch Changes